requirement.txt
aiosmtplib==1.1.7
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
files==1.1.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.4
messages==0.8.0
Pillow==9.4.0
pypdf==3.5.1
PyPDF2==3.0.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
validus==0.3.0
